+++ src/std_nodes/condition.rs
@@ -77,7 +77,7 @@ impl<T: Send + Sync + 'static> Node<T> for Condition<T>
 			id: self.id,
 			num_children: 0,
 			children: Vec::new(),
-			status: self.status,
+			status: self.status() as i32,
 			type_name: "Condition".to_string(),
 		}
 	}
